@@ -4,14 +4,14 @@
     "type": "asgard-module",
     "license": "MIT",
     "require": {
-        "php": "^7.1.3",
+        "php": "^8.1",
         "composer/installers": "~1.0",
-        "imagina/core-module": "v8.x-dev",
-        "imagina/requestable-module": "v8.x-dev"
+        "imagina/core-module": "v10.x-dev",
+        "imagina/requestable-module": "v10.x-dev"
     },
     "require-dev": {
-        "phpunit/phpunit": "~7.0",
-        "orchestra/testbench": "3.8.*"
+        "orchestra/testbench": "^8.5",
+        "phpunit/phpunit": "^10"
     },
     "autoload-dev": {
         "psr-4": {
@@ -29,7 +29,7 @@
     ],
     "extra": {
         "branch-alias": {
-            "dev-v8.x": "8.0.x-dev"
+            "dev-v10.x": "10.0.x-dev"
         }
     },
     "autoload": {
