@@ -1,8 +1,8 @@
 <?php
 
 return [
-  "admin" => [],
-  "panel" => [
-    "icredit_cms_main_wallet"
-  ],
+    'admin' => [],
+    'panel' => [
+        'icredit_cms_main_wallet',
+    ],
 ];
