@@ -1,15 +1,15 @@
 <?php
 
 return [
-     'formFields' => [
+    'formFields' => [
         'title' => [
             'value' => null,
             'name' => 'title',
             'type' => 'input',
             'isTranslatable' => true,
             'props' => [
-               'label' => 'icommerce::common.title'
-            ]
+                'label' => 'icommerce::common.title',
+            ],
         ],
         'description' => [
             'value' => null,
@@ -20,7 +20,7 @@ return [
                 'label' => 'icommerce::common.description',
                 'type' => 'textarea',
                 'rows' => 3,
-            ]
+            ],
         ],
         'status' => [
             'value' => 0,
@@ -34,35 +34,35 @@ return [
                 'hideDropdownIcon' => true,
                 'newValueMode' => 'add-unique',
                 'options' => [
-                    ['label' => 'Activo','value' => 1],
-                    ['label' => 'Inactivo','value' => 0],
-                ]
-            ]
+                    ['label' => 'Activo', 'value' => 1],
+                    ['label' => 'Inactivo', 'value' => 0],
+                ],
+            ],
         ],
         'mainimage' => [
-            'value' => (object)[],
+            'value' => (object) [],
             'name' => 'mediasSingle',
             'type' => 'media',
             'props' => [
                 'label' => 'Image',
                 'zone' => 'mainimage',
                 'entity' => "Modules\Icommerce\Entities\PaymentMethod",
-                'entityId' => null
-            ]
+                'entityId' => null,
+            ],
         ],
         'init' => [
             'value' => 'Modules\Icredit\Http\Controllers\Api\PaymentApiController',
             'name' => 'init',
-            'isFakeField' => true
+            'isFakeField' => true,
         ],
         'minimunAmount' => [
-          'value' => null,
-          'name' => 'minimunAmount',
-          'isFakeField' => true,
-          'type' => 'input',
-          'props' => [
-                'label' => 'icommerce::common.minimum Amount'
-          ]
+            'value' => null,
+            'name' => 'minimunAmount',
+            'isFakeField' => true,
+            'type' => 'input',
+            'props' => [
+                'label' => 'icommerce::common.minimum Amount',
+            ],
         ],
-    ]
+    ],
 ];
