@@ -1,14 +1,14 @@
 <?php
 
 return [
-  'icredit.credits' => [
-    'manage' => 'icredit::credits.manage resource',
-    'index' => 'icredit::credits.list resource',
-    'create' => 'icredit::credits.create resource',
-    'edit' => 'icredit::credits.edit resource',
-    'destroy' => 'icredit::credits.destroy resource',
-    'restore' => 'icredit::credits.restore resource',
-  ],
-// append
+    'icredit.credits' => [
+        'manage' => 'icredit::credits.manage resource',
+        'index' => 'icredit::credits.list resource',
+        'create' => 'icredit::credits.create resource',
+        'edit' => 'icredit::credits.edit resource',
+        'destroy' => 'icredit::credits.destroy resource',
+        'restore' => 'icredit::credits.restore resource',
+    ],
+    // append
 
 ];
