@@ -1,32 +1,26 @@
 <?php
 
-use Illuminate\Support\Facades\Schema;
-use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 
-class IcreditChangeCustomerIdInCreditTable extends Migration
+return new class extends Migration
 {
     /**
      * Run the migrations.
-     *
-     * @return void
      */
-    public function up()
+    public function up(): void
     {
-      Schema::table('icredit__credits', function (Blueprint $table) {
-
-        $table->integer('customer_id')->unsigned()->nullable()->change();
-
-      });
+        Schema::table('icredit__credits', function (Blueprint $table) {
+            $table->integer('customer_id')->unsigned()->nullable()->change();
+        });
     }
 
     /**
      * Reverse the migrations.
-     *
-     * @return void
      */
-    public function down()
+    public function down(): void
     {
         //
     }
-}
+};
