@@ -2,32 +2,29 @@
 
 namespace Modules\Icredit\Database\Seeders;
 
-use Illuminate\Database\Seeder;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Seeder;
 
 class IcreditModuleTableSeeder extends Seeder
 {
-  /**
-   * Run the database seeds.
-   *
-   * @return void
-   */
-  public function run()
-  {
-    Model::unguard();
-  
+    /**
+     * Run the database seeds.
+     */
+    public function run()
+    {
+        Model::unguard();
 
-    $columns = [
-      ["config" => "cms-pages", "name" => "cms_pages"],
-      ["config" => "cms-sidebar", "name" => "cms_sidebar"],
-      ["config" => "config", "name" => "config"],
-      ["config" => "crud-fields", "name" => "crud_fields"],
-      ["config" => "permissions", "name" => "permissions"],
-      ["config" => "settings-fields", "name" => "settings"],
-    ];
-  
-    $moduleRegisterService = app("Modules\Isite\Services\RegisterModuleService");
-  
-    $moduleRegisterService->registerModule("icredit", $columns, 1);
-  }
+        $columns = [
+            ['config' => 'cms-pages', 'name' => 'cms_pages'],
+            ['config' => 'cms-sidebar', 'name' => 'cms_sidebar'],
+            ['config' => 'config', 'name' => 'config'],
+            ['config' => 'crud-fields', 'name' => 'crud_fields'],
+            ['config' => 'permissions', 'name' => 'permissions'],
+            ['config' => 'settings-fields', 'name' => 'settings'],
+        ];
+
+        $moduleRegisterService = app("Modules\Isite\Services\RegisterModuleService");
+
+        $moduleRegisterService->registerModule('icredit', $columns, 1);
+    }
 }
