@@ -4,12 +4,13 @@ namespace Modules\Icredit\Entities;
 
 /**
  * Class Status
- * @package Modules\icommerce\Entities
  */
 class Status
 {
     const PENDING = 1;
+
     const APPROVED = 2;
+
     const CANCELED = 3;
 
     /**
@@ -28,19 +29,16 @@ class Status
 
     /**
      * Get the available statuses
-     * @return array
      */
-    public function lists()
+    public function lists(): array
     {
         return $this->statuses;
     }
 
     /**
      * Get the post status
-     * @param int $statusId
-     * @return string
      */
-    public function get($statusId)
+    public function get(int $statusId): string
     {
         if (isset($this->statuses[$statusId])) {
             return $this->statuses[$statusId];
