@@ -2,13 +2,12 @@
 
 namespace Modules\Icredit\Http\Controllers\Admin;
 
-use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Modules\Core\Http\Controllers\Admin\AdminBaseController;
 use Modules\Icredit\Entities\Credit;
 use Modules\Icredit\Http\Requests\CreateCreditRequest;
 use Modules\Icredit\Http\Requests\UpdateCreditRequest;
 use Modules\Icredit\Repositories\CreditRepository;
-use Modules\Core\Http\Controllers\Admin\AdminBaseController;
 
 class CreditController extends AdminBaseController
 {
@@ -26,10 +25,8 @@ class CreditController extends AdminBaseController
 
     /**
      * Display a listing of the resource.
-     *
-     * @return Response
      */
-    public function index()
+    public function index(): Response
     {
         //$credits = $this->credit->all();
 
@@ -38,21 +35,16 @@ class CreditController extends AdminBaseController
 
     /**
      * Show the form for creating a new resource.
-     *
-     * @return Response
      */
-    public function create()
+    public function create(): Response
     {
         return view('icredit::admin.credits.create');
     }
 
     /**
      * Store a newly created resource in storage.
-     *
-     * @param  CreateCreditRequest $request
-     * @return Response
      */
-    public function store(CreateCreditRequest $request)
+    public function store(CreateCreditRequest $request): Response
     {
         $this->credit->create($request->all());
 
@@ -62,23 +54,16 @@ class CreditController extends AdminBaseController
 
     /**
      * Show the form for editing the specified resource.
-     *
-     * @param  Credit $credit
-     * @return Response
      */
-    public function edit(Credit $credit)
+    public function edit(Credit $credit): Response
     {
         return view('icredit::admin.credits.edit', compact('credit'));
     }
 
     /**
      * Update the specified resource in storage.
-     *
-     * @param  Credit $credit
-     * @param  UpdateCreditRequest $request
-     * @return Response
      */
-    public function update(Credit $credit, UpdateCreditRequest $request)
+    public function update(Credit $credit, UpdateCreditRequest $request): Response
     {
         $this->credit->update($credit, $request->all());
 
@@ -88,11 +73,8 @@ class CreditController extends AdminBaseController
 
     /**
      * Remove the specified resource from storage.
-     *
-     * @param  Credit $credit
-     * @return Response
      */
-    public function destroy(Credit $credit)
+    public function destroy(Credit $credit): Response
     {
         $this->credit->destroy($credit);
 
