@@ -1,9 +1,8 @@
 <?php
+
 use Illuminate\Routing\Router;
 
-$router->group(['prefix' => '/icredit/v1'], function (Router $router) {
-
-   
+Route::prefix('/icredit/v1')->group(function (Router $router) {
     // Base API Routes
     $router->apiCrud([
         'module' => 'icredit',
@@ -12,16 +11,13 @@ $router->group(['prefix' => '/icredit/v1'], function (Router $router) {
         //'middleware' =>  ['create' => [],'update' => [],'delete' => [],'restore' => []]
     ]);
 
-    // withdrawalFunds 
+    // withdrawalFunds
     $router->post('/withdrawal-funds', [
         'as' => 'icredit.api.credit.withdrawalFunds',
         'uses' => 'CreditApiController@withdrawalFunds',
-        'middleware' => ['auth:api']
+        'middleware' => ['auth:api'],
     ]);
 
-    
     //======  PaymentMethod
-    require('ApiRoutes/paymentRoutes.php');
-
-
+    require 'ApiRoutes/paymentRoutes.php';
 });
