@@ -2,11 +2,9 @@
 
 use Illuminate\Routing\Router;
 
-$router->group(['prefix'=>'icredit'],function (Router $router){
-       
+Route::prefix('icredit')->group(function (Router $router) {
     $router->get('payment/{eUrl}', [
         'as' => 'icredit.payment.index',
         'uses' => 'PublicController@paymentIndex',
     ]);
-       
 });
