@@ -2,16 +2,16 @@
 
 namespace Modules\Icredit\Providers;
 
-use Illuminate\Database\Eloquent\Factory as EloquentFactory;
 use Illuminate\Support\ServiceProvider;
-use Modules\Core\Traits\CanPublishConfiguration;
 use Modules\Core\Events\BuildingSidebar;
 use Modules\Core\Events\LoadingBackendTranslations;
+use Modules\Core\Traits\CanPublishConfiguration;
 use Modules\Icredit\Listeners\RegisterIcreditSidebar;
 
 class IcreditServiceProvider extends ServiceProvider
 {
     use CanPublishConfiguration;
+
     /**
      * Indicates if loading of the provider is deferred.
      *
@@ -21,10 +21,8 @@ class IcreditServiceProvider extends ServiceProvider
 
     /**
      * Register the service provider.
-     *
-     * @return void
      */
-    public function register()
+    public function register(): void
     {
         $this->registerBindings();
         $this->app['events']->listen(BuildingSidebar::class, RegisterIcreditSidebar::class);
@@ -32,34 +30,29 @@ class IcreditServiceProvider extends ServiceProvider
         $this->app['events']->listen(LoadingBackendTranslations::class, function (LoadingBackendTranslations $event) {
             $event->load('credits', Arr::dot(trans('icredit::credits')));
             // append translations
-
         });
-
-
     }
 
-    public function boot()
+    public function boot(): void
     {
         $this->publishConfig('icredit', 'config');
         $this->publishConfig('icredit', 'crud-fields');
 
-      $this->mergeConfigFrom($this->getModuleConfigFilePath('icredit', 'settings'), "asgard.icredit.settings");
-      $this->mergeConfigFrom($this->getModuleConfigFilePath('icredit', 'settings-fields'), "asgard.icredit.settings-fields");
-      $this->mergeConfigFrom($this->getModuleConfigFilePath('icredit', 'permissions'), "asgard.icredit.permissions");
-      $this->mergeConfigFrom($this->getModuleConfigFilePath('icredit', 'cmsPages'), "asgard.icredit.cmsPages");
-      $this->mergeConfigFrom($this->getModuleConfigFilePath('icredit', 'cmsSidebar'), "asgard.icredit.cmsSidebar");
+        $this->mergeConfigFrom($this->getModuleConfigFilePath('icredit', 'settings'), 'asgard.icredit.settings');
+        $this->mergeConfigFrom($this->getModuleConfigFilePath('icredit', 'settings-fields'), 'asgard.icredit.settings-fields');
+        $this->mergeConfigFrom($this->getModuleConfigFilePath('icredit', 'permissions'), 'asgard.icredit.permissions');
+        $this->mergeConfigFrom($this->getModuleConfigFilePath('icredit', 'cmsPages'), 'asgard.icredit.cmsPages');
+        $this->mergeConfigFrom($this->getModuleConfigFilePath('icredit', 'cmsSidebar'), 'asgard.icredit.cmsSidebar');
 
         //$this->loadMigrationsFrom(__DIR__ . '/../Database/Migrations');
     }
 
     /**
      * Get the services provided by the provider.
-     *
-     * @return array
      */
-    public function provides()
+    public function provides(): array
     {
-        return array();
+        return [];
     }
 
     private function registerBindings()
@@ -76,9 +69,6 @@ class IcreditServiceProvider extends ServiceProvider
                 return new \Modules\Icredit\Repositories\Cache\CacheCreditDecorator($repository);
             }
         );
-// add bindings
-
+        // add bindings
     }
-
-
 }
