@@ -8,31 +8,23 @@ class RouteServiceProvider extends CoreRoutingServiceProvider
 {
     /**
      * The root namespace to assume when generating URLs to actions.
+     *
      * @var string
      */
     protected $namespace = 'Modules\Icredit\Http\Controllers';
 
-    /**
-     * @return string
-     */
-    protected function getFrontendRoute()
+    protected function getFrontendRoute(): string
     {
-        return __DIR__ . '/../Http/frontendRoutes.php';
+        return __DIR__.'/../Http/frontendRoutes.php';
     }
 
-    /**
-     * @return string
-     */
-    protected function getBackendRoute()
+    protected function getBackendRoute(): string
     {
-        return __DIR__ . '/../Http/backendRoutes.php';
+        return __DIR__.'/../Http/backendRoutes.php';
     }
 
-    /**
-     * @return string
-     */
-    protected function getApiRoute()
+    protected function getApiRoute(): string
     {
-        return __DIR__ . '/../Http/apiRoutes.php';
+        return __DIR__.'/../Http/apiRoutes.php';
     }
 }
