@@ -6,9 +6,7 @@ use Modules\Core\Repositories\BaseRepository;
 
 interface CreditRepository extends BaseRepository
 {
+    public function calculate($parameters, $conf);
 
-    public function calculate($parameters,$conf);
-    
     //public function amount($params);
-    
 }
