@@ -3,6 +3,6 @@
 return [
     'setup-wallet' => [
         'title' => 'Set up your wallet',
-        'description' => 'You must configure your payment method in order to receive payments'
+        'description' => 'You must configure your payment method in order to receive payments',
     ],
 ];
