@@ -1,18 +1,17 @@
 <?php
 
 return [
-  "sidebar" => [
-    "adminWallet" => "Wallet"
-  ],
-  "label" => [
-    "wallet" => "Wallet",
-    "amountAvailable" => "Amount available",
-    "withdrawal" => "withdrawal"
-  ],
-  "message" => [
-    "notAmountAvailable" => "No amount available",
-    "invalidAmount" => "The value entered for withdrawal is not valid",
-    "showPanel" => "Show Panel"
-  ]
+    'sidebar' => [
+        'adminWallet' => 'Wallet',
+    ],
+    'label' => [
+        'wallet' => 'Wallet',
+        'amountAvailable' => 'Amount available',
+        'withdrawal' => 'withdrawal',
+    ],
+    'message' => [
+        'notAmountAvailable' => 'No amount available',
+        'invalidAmount' => 'The value entered for withdrawal is not valid',
+        'showPanel' => 'Show Panel',
+    ],
 ];
-
