@@ -1,9 +1,9 @@
 <?php
+
 return [
-  "name" => "Wallet",
-  'validation' => [
-    'no credit' => 'Insufficient credit'
-  ],
-  
+    'name' => 'Wallet',
+    'validation' => [
+        'no credit' => 'Insufficient credit',
+    ],
 
 ];
