@@ -3,6 +3,6 @@
 return [
     'setup-wallet' => [
         'title' => 'Configura tu billetera',
-        'description' => 'Debes configurar tu metodo de cobro para poder recibir los pagos'
+        'description' => 'Debes configurar tu metodo de cobro para poder recibir los pagos',
     ],
 ];
