@@ -1,18 +1,17 @@
 <?php
 
 return [
-  "sidebar" => [
-    "adminWallet" => "Mi Billetera"
-  ],
-  "label" => [
-    "wallet" => "Billetera",
-    "amountAvailable" => "Cantidad disponible",
-    "withdrawal" => "Nuevo retiro"
-  ],
-  "message" => [
-    "notAmountAvailable" => "Sin saldo disponible",
-    "invalidAmount" => "El valor ingresado para el retiro no es valido",
-    "showPanel" => "Ver mi Panel"
-  ]
+    'sidebar' => [
+        'adminWallet' => 'Mi Billetera',
+    ],
+    'label' => [
+        'wallet' => 'Billetera',
+        'amountAvailable' => 'Cantidad disponible',
+        'withdrawal' => 'Nuevo retiro',
+    ],
+    'message' => [
+        'notAmountAvailable' => 'Sin saldo disponible',
+        'invalidAmount' => 'El valor ingresado para el retiro no es valido',
+        'showPanel' => 'Ver mi Panel',
+    ],
 ];
-
