@@ -1,7 +1,8 @@
 <?php
+
 return [
-  "name" => "Billetera",
-  'validation' => [
-    'no credit' => 'Crédito insuficiente (:creditUser)'
-  ],
+    'name' => 'Billetera',
+    'validation' => [
+        'no credit' => 'Crédito insuficiente (:creditUser)',
+    ],
 ];
