@@ -2,31 +2,27 @@
 
 namespace Modules\Icredit\Services;
 
+use Illuminate\Http\Request;
 use Modules\Icredit\Entities\Credit;
 use Modules\Icredit\Repositories\CreditRepository;
-use Illuminate\Http\Request;
 
 class CreditService
 {
-
     public function __construct(
 
         CreditRepository $credit
-    )
-    {
+    ) {
         $this->credit = $credit;
-
     }
 
     /**
      * CREATE A ITEM
      *
-     * @param Request $request
+     * @param  Request  $request
      * @return mixed
      */
     public function create($data)
     {
-
         $needTobeCreated = true;
         if (isset($data['creditId'])) {
             $credit = $this->credit->find($data['creditId']);
@@ -34,28 +30,25 @@ class CreditService
         } elseif (isset($data['credit']->id)) {
             $credit = $data['credit'];
             $needTobeCreated = false;
-            if(!isset($credit->id)){
+            if (! isset($credit->id)) {
                 $needTobeCreated = true;
             }
         }
 
-        if($needTobeCreated) {
-
+        if ($needTobeCreated) {
             $creditData = [
-                "amount" => $data["amount"] ?? 0,
-                "customer_id" => $data["customerId"] ?? null,
-                "description" => $data["description"] ?? "",
-                "status" => $data["status"] ?? 1,
-                "related_id" => $data["relatedId"] ?? null,
-                "related_type" => $data["relatedType"] ?? null
+                'amount' => $data['amount'] ?? 0,
+                'customer_id' => $data['customerId'] ?? null,
+                'description' => $data['description'] ?? '',
+                'status' => $data['status'] ?? 1,
+                'related_id' => $data['relatedId'] ?? null,
+                'related_type' => $data['relatedType'] ?? null,
             ];
-
 
             //Create credit
             $credit = $this->credit->create($creditData);
-
         }
+
         return $credit;
     }
-
 }
