@@ -6,26 +6,19 @@ use Modules\Core\Icrud\Transformers\CrudResource;
 
 class CreditTransformer extends CrudResource
 {
-	
-	/**
-	  * Method to merge values with response
-	  *
-	  * @return array
-	  */
-	public function modelAttributes($request)
-	{
-		
-		$filter = json_decode($request->filter);
+    /**
+     * Method to merge values with response
+     */
+    public function modelAttributes($request)
+    {
+        $filter = json_decode($request->filter);
 
-		
-		if(isset($filter) && isset($filter->amountAvailable)){
-			return [];
-		}else{
-			return [
-		      'statusName' => $this->statusName,
-		    ]; 	
-		}
-		   
-	}
-    
+        if (isset($filter) && isset($filter->amountAvailable)) {
+            return [];
+        } else {
+            return [
+                'statusName' => $this->statusName,
+            ];
+        }
+    }
 }
